refactor(main): 메인 페이지 책임 분리
